Moved AnimatedImageCache to a separate file
